feat

+++ watch-tests.js
@@ -2,10 +2,9 @@ const { spawn } = require("child_process");
 const fs = require("fs");
 
 const triggerFile = ".trigger";
+const resultFile = ".test-result";
 
-if (!fs.existsSync(triggerFile)) {
-  fs.writeFileSync(triggerFile, "");
-}
+if (!fs.existsSync(triggerFile)) fs.writeFileSync(triggerFile, "");
 
 console.log("[TEST RUNNER] Aguardando o clique no botão RUN TESTS...");
 
@@ -13,9 +12,22 @@ fs.watch(triggerFile, (eventType) => {
   if (eventType === "change") {
     console.log("\n[TEST RUNNER] Disparando testes...\n");
 
-    spawn("npx", ["ng", "test"], {
+    fs.writeFileSync(resultFile, "RUNNING");
+
+    const testProcess = spawn("npx", ["ng", "test"], {
       stdio: "inherit",
       shell: true,
+    });
+
+    testProcess.on("close", (code) => {
+      const status = code === 0 ? "PASS" : "FAIL";
+      fs.writeFileSync(resultFile, status);
+
+      if (status === "PASS") {
+        console.log("\nTODOS OS TESTES PASSARAM!");
+      } else {
+        console.log("\nALGUNS TESTES FALHARAM.");
+      }
     });
   }
 });
